Refactor project repositories and remove travis
