--- .travis.yml
@@ -1,9 +1,0 @@
-language: python
-python:
-  - "3.7"
-
-install:
-  - pip install -r requirements.txt
-
-script:
-  - pytest
