+++ pyproject.toml
@@ -32,18 +32,14 @@ dependencies = [
 "Homepage" = "https://github.com/cnescatlab/polyspace-report2excel"
 "Bug Reports" = "https://github.com/cnescatlab/polyspace-report2excel/issues"
 
-# This is configuration specific to the `setuptools` build backend.
-# If you are using a different build backend, you will need to change this.
+[project.scripts]
+p2e = "p2e:main"
+
+
 [tool.setuptools]
-# If there are data files included in your packages that need to be
-# installed, specify them here.
 package-data = {}
 py-modules = []
 
 [build-system]
 requires = ["setuptools>=61.0", "wheel"]
 build-backend = "setuptools.build_meta"
-
-# Entry points
-[project.scripts]
-export-all = "polyspace-report2excel.export-all"
