--- requirements.txt
@@ -1,3 +1,0 @@
-XlsxWriter
-html2text
-striprtf
