+++ src/p2e/__init__.py
@@ -1,0 +1,23 @@
+"""
+Modules:
+export-all:
+    export Polyspace report (html or RTF) to excel file
+exportcsv:
+    write to csv file
+exportxlsx:
+    write to excel file
+HTMLReader:
+    read HTML file and extract arrays from it
+RTFReader:
+    read RTF file and extract arrays from it
+main:
+    entrypoint for using this module
+"""
+
+def main():
+    import sys
+    import main
+    sys.exit(export-all.main())
+
+if __name__ == "__main__":
+    main()
